k64ay/events_app/regee/requirements.txt
Django==4.1.5
python-dotenv==0.21.0
Pillow==9.4.0
selenium==4.8.3
python-dateutil==2.8.2